backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
pydantic[email]>=2.4.2
pydantic-settings>=2.0.3
bcrypt>=4.0.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
